Доработан генетический алгоритм

+++ src/algorithms/algorithm.rs
@@ -6,6 +6,6 @@ pub trait OptimizationAlgorithm {
 
 pub trait OptimizationAlgorithmBuilder {
     fn iters_count(self, iters_count: u64) -> Self;
-    fn actors_count(self, actors_count: u64) -> Self;
+    fn actors_count(self, actors_count: usize) -> Self;
     fn solutions_count(self, solutions_count: u64) -> Self;
 }
